Login, Perfil Adopter Rectification

+++ PetCareFront/PetCare/src/app/adoptante/adoptante.component.ts
@@ -23,7 +23,7 @@ export class AdoptanteComponent implements OnInit{
 
   formatDateString(dateString: string): string {
     const date = new Date(dateString);
-    const options: Intl.DateTimeFormatOptions = { year: 'numeric', month: 'long', day: 'numeric' };
+    const options: Intl.DateTimeFormatOptions = { year: 'numeric', month: 'numeric', day: 'numeric' };
     return date.toLocaleDateString('es-ES', options);
   }
